Moving the spring kafka example to a separate module
* So that we can have multiple types of examples in separate
  modules
