+++ kafka-spring/src/main/java/com/tib/kafka/spring/KafkaPlayground.java
@@ -1,4 +1,4 @@
-package com.tib.kafka;
+package com.tib.kafka.spring;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
