+++ kafka-spring/src/main/java/com/tib/kafka/spring/KafkaPlaygroundConstants.java
@@ -1,4 +1,4 @@
-package com.tib.kafka;
+package com.tib.kafka.spring;
 
 public class KafkaPlaygroundConstants {
     public static final String TOPIC_NAME = "mytest";
